Fixed hitboxes and locations

+++ core/src/com/mygdx/game/Model/Location.java
@@ -3,11 +3,11 @@ package com.mygdx.game.Model;
 public class Location {
 
     private String locationName;
-    private int xPos;
-    private int yPos;
+    private float xPos;
+    private float yPos;
 
     //Constructor
-    public Location(String locationName, int xPos, int yPos) {
+    public Location(String locationName, float xPos, float yPos) {
         this.locationName = locationName;
         this.xPos = xPos;
         this.yPos = yPos;
